Input Data for Well Capacity Prediction

A reduced dataset for well capacity prediction created with R script in /data-raw/prepare_model_data.R

Usage

model_data_reduced

Format

A data.frame with 6308 rows and 27 variables:

Qs_rel

specific capacity of well relative to operational start condition, output

well_id

well id, for info

well_age_years

years since operationa start, input, numeric

construction_year

year of well construction

screen_material

screen material

diameter

well diameter (mm)

drilling_method

drilling_method

admissible_discharge

allowed pumping rate

operational_start.Qs

initial Qs at construction

aquifer_coverage

confined / unconfined

W_static.sd

standard deviation of static water level

surface_water.distance

distance to surface water

n_rehab

number of well rehabilitations

time_since_rehab_years

time since last well rehabilitation in years

volume_m3_d.mean

mean daily abstraction volume (m3)

volume_m3_d.cv

coefficient of variation of daily abstraction volume (m3)

quality.EC

water quality: electical conductivity (us/cm)

quality.D0

water quality: dissolved oxygen (mg/l)

quality.Temp

water quality: temperature (C)

quality.pH

water quality: pH

quality.Redox

water quality: electical conductivity (us/cm)

quality.Fe_tot

water quality: dissolved oxygen (mg/l)

quality.Mn

water quality: Mn (mg/l)

quality.NO3

water quality: NO3 (mg/l)

quality.PO4

water quality: PO4 (mg/l)

quality.SO4

water quality: SO4 (mg/l)

quality.TSS

water quality: Total Suspended Solids (mg/l)

Examples

well_ids <- unique(rehabs$well_id[1:20])
idx <- model_data_reduced$well_id %in% well_ids
tibble::as_tibble(model_data_reduced[idx, ])
#> # A tibble: 43 × 27
#>    Qs_rel well_id well_age_years const…¹ scree…² diame…³ drill…⁴ admis…⁵ opera…⁶
#>     <dbl>   <int>          <dbl>   <dbl> <fct>     <dbl> <fct>     <dbl>   <dbl>
#>  1 100       9690           0       1991 d94670…     350 bbdca1…    50      50.2
#>  2  70.5     9690           3.79    1991 d94670…     350 bbdca1…    50      50.2
#>  3   9.68    9690          13.2     1991 d94670…     350 bbdca1…    50      50.2
#>  4  44.6     9690          13.4     1991 d94670…     350 bbdca1…    50      50.2
#>  5 100       3328           0       1995 93242c…     400 bbdca1…    46.8    25.4
#>  6  76.7     3328           5.92    1995 93242c…     400 bbdca1…    46.8    25.4
#>  7  86.5     3328           5.96    1995 93242c…     400 bbdca1…    46.8    25.4
#>  8  39.4     3328          14.5     1995 93242c…     400 bbdca1…    46.8    25.4
#>  9  60.3     3328          14.6     1995 93242c…     400 bbdca1…    46.8    25.4
#> 10  19.9     3328          17.2     1995 93242c…     400 bbdca1…    46.8    25.4
#> # … with 33 more rows, 18 more variables: aquifer_coverage <fct>,
#> #   W_static.sd <dbl>, surface_water.distance <fct>, n_rehab <int>,
#> #   time_since_rehab_years <dbl>, volume_m3_d.mean <dbl>, volume_m3_d.cv <dbl>,
#> #   quality.EC <dbl>, quality.DO <dbl>, quality.Temp <dbl>, quality.pH <dbl>,
#> #   quality.Redox <dbl>, quality.Fe_tot <dbl>, quality.Mn <dbl>,
#> #   quality.NO3 <dbl>, quality.PO4 <dbl>, quality.SO4 <dbl>, quality.TSS <dbl>,
#> #   and abbreviated variable names ¹​construction_year, ²​screen_material, …